requirements.txt
mock==4.0.3
pycrypto==2.6.1
pytest==6.2.1
Sphinx==3.4.3
sphinx-rtd-theme==0.5.1
